tests/wpt/web-platform-tests/performance-timeline/tentative/include-frames-one-same-origin-child.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
</body>
<script>
promise_test(() => {
    return new Promise(resolve => {
        performance.clearResourceTimings()

        // Create child iframe.
        const childFrame = document.createElement('iframe')
        childFrame.src = "../resources/child-frame.html"
        document.body.appendChild(childFrame)

        // On-load event handler to assert size of entries.
        childFrame.addEventListener("load", () => {
            const entries = performance.getEntries(true)
            const navigationEntries = performance.getEntriesByType("navigation", true)
            const markedEntries = performance.getEntriesByName("entry-name", undefined, true)

            // 3 entries for parent, 2 for child.
            assert_equals(entries.length, 5)

            // 1 entry for parent, 1 for child.
            assert_equals(navigationEntries.length, 2)

            // 1 entry for child.
            assert_equals(markedEntries.length, 1)

            resolve()
        })
    })
}, "GetEntries of a parent Frame with one Same-Origin child");
</script>
